Stefirta

+++ src/main/java/appLogic/activity/impl/WorkActivity.java
@@ -1,15 +1,15 @@
 package appLogic.activity.impl; //s244813
 
 public class WorkActivity extends Activity {
-    private final int expectedHours;
+    private final double expectedHours;
 
     public WorkActivity(String name, String description, String summary,
-                        int startWeek, int endWeek, int startYear, int endYear, int expectedHours) {
+                        int startWeek, int endWeek, int startYear, int endYear, double expectedHours) {
         super(name, description, summary, startWeek, endWeek, startYear, endYear, null);
         this.expectedHours = expectedHours;
     }
 
-    public int getExpectedHours() {
+    public double getExpectedHours() {
         return expectedHours;
     }
 }
